admin detection feature

+++ CMS/admin/includes/view_all_users.php
@@ -30,15 +30,6 @@
         echo "<td>$user_email</td>";
         echo "<td>$user_role</td>";
         
-
-//        $query = "SELECT * FROM posts WHERE post_id = $comment_post_id";
-//        $select_post_id_query = mysqli_query($connection, $query);
-//        while($row = mysqli_fetch_assoc($select_post_id_query)){
-//            $post_id = $row['post_id'];
-//            $post_title = $row['post_title'];
-//            echo "<td><a href='../post.php?p_id=$post_id'>$post_title</a></td>";
-//        }
-        
         echo "<td><a href='users.php?change_to_admin={$user_id}'>Admin</a></td>";
         echo "<td><a href='users.php?change_to_sub={$user_id}'>Subscriber</a></td>";
         echo "<td><a href='users.php?source=edit_user&edit_user={$user_id}'>Edit</a></td>";
@@ -67,10 +58,15 @@ if(isset($_GET['change_to_sub'])){
 
 
 if(isset($_GET['delete'])){
-    $the_user_id = $_GET['delete'];
-    $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
-    $delete_query = mysqli_query($connection, $query);
-    header("Location:users.php");
+    if(isset($_SESSION['role'])){
+        
+        if($_SESSION['role'] == 'admin'){
+            $the_user_id = mysqli_real_escape_string($connection, $_GET['delete']);
+            $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
+            $delete_query = mysqli_query($connection, $query);
+            header("Location:users.php");
+        }
+    }
 }
 
 ?>
